pyproject.toml
[project]
name = "tableau_langchain"
version = "0.1.0"
description = "A community project to stage, test and build Tableau tools for Agents built with the Langgraph framework"
authors = [
    { name = "Stephen Price" },
    { name = "Joe Constantino" },
    { name = "Joseph Fluckiger" },
    { name = "Antoine Issaly" }
]
requires-python = ">=3.12.2"
license = "MIT"
readme = "README.md"

dependencies = [
    "build==1.2.2",
    "ipython==9.0.2",
    "langchain==0.3.21",
    "langgraph>=0.4.20",
    "langgraph-sdk>=0.2.0",
    "langsmith>=0.3.45",
    "langchain-anthropic==0.3.10",
    "langchain-openai==0.3.11",
    "langchain-pinecone==0.2.3",
    "langchain-tableau",
    "langchain-community==0.3.20",
    "pinecone-client[grpc]==6.0.0",
    "python-dotenv==1.1.0",
    "slack-sdk==3.35.0",
    "tavily-python==0.5.3",
    "twine==6.1.0"
]

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["experimental"]
